requirements-pip.txt
Cython==0.19.2
Flask==0.10.1
Jinja2==2.7.1
MDP==3.3
MarkupSafe==0.18
Pillow==2.2.1
PyYAML==3.10
Pygments==1.6
SQLAlchemy==0.8.3
SimpleCV==1.3
Sphinx==1.1.3
Werkzeug==0.9.4
astropy==0.2.5
beautifulsoup4==4.3.1
bitarray==0.8.1
colorama==0.2.7
distribute==0.6.45
docutils==0.11
greenlet==0.4.1
ipython==1.1.0
itsdangerous==0.23
llvmmath==0.1
llvmpy==0.12.0
lxml==3.2.3
matplotlib==1.3.1
netCDF4==1.0.6
networkx==1.8.1
nose==1.3.0
numba==0.11.0
numexpr==2.2.2
numpy==1.7.1
openpyxl==1.6.2
pandas==0.12.0
patsy==0.2.1
ply==3.4
psutil==1.1.2
pycosat==0.6.0
pycparser==2.09.1
pycrypto==2.6.1
pyflakes==0.7.3
pyparsing==2.0.1
python-dateutil==2.1
pytz==2013b
pyzmq==2.2.0.1
requests==1.2.3
scikit-image==0.9.3
scikit-learn==0.14.1
scikits-image==0.7.1
scipy==0.13.0
six==1.4.1
statsmodels==0.5.0
sympy==0.7.3
tables==3.0.0
tornado==3.1.1
xlrd==0.9.2
